+++ CHANGELOG.md
@@ -9,6 +9,7 @@
 * New block on return in editor @robgietema
 * Added 404 page on page not found @robgietema
 * Custom tiles support @sneridagh
+* Add full register/password reset views @sneridagh
 
 ### Changes